move derivative type check from class to function
This patch moves the type check 'derivative target is an `Argument`' from class
`_Derivative` to function `derivative`. The idea behind this change is that
function `derivative` is public and should handle invalid arguments with proper
exceptions, while class `_Derivative` is private and we don't make that kind of
errors internally.