bump minimal numpy version to 1.17
This patch bumps the minimum required Numpy version from 1.16 to 1.17, which is
the first version that enables Numpy's array function protocol by default.
While Numpy 1.16 does support this protocol, it requires the
NUMPY_EXPERIMENTAL_ARRAY_FUNCTION environment to be set, which is more effort
that it is worth to support an old version.

Additionally, this patch changes the testing matrix by lowering the Python
version for which Numpy 1.17 is tested from 3.8 to 3.6, which is the highest
version for which pypi hosts precompiled wheels.