remove LRU nature of lru_cache
This patch removes the LRU aspect of lru_cache in favour of a new restriction
demending that cached functions have at least one array argument to trigger
cache deletion. As this reduces the cache to a special kind of weak key
dictionary, which cannot grow beyond the set of live arguments, a size cap is
no longer required to prevent leaks.

While the new character merits a name change to arraycache or buffercache, the
old name is maintained to minimize disruption until a decision is made about
the long term use of this construct.