fix sample.eval_integrals_sparse
This patch fixes a bug that was introduced in commit a202366 of #653, which
introduced evaluable.eval_sparse and deprecated sample.eval_integrals_sparse.
In doing so, the method was erroneously changed to dispatch to
function.eval_sparse, a location that had been temporarily considered but was
ultimately rejected. The bug was not noticed because all unit tests were
updated to use the new method directly; two tests are added to correct this.