remove sphinx version cap
With the release of docutils 0.18, older versions of sphinx such as v3.0.4
result in a sphinx.errors.SphinxWarning (node class 'meta' is already
registered, its visitors will be overridden). The github workflow capped the
Sphinx version at below 3.1 for reasons that do not appear relevant anymore.
Releaving the cap fixes the docutils issue.