let evaluable.Reciprocal have dtype of argument
`evaluable.Reciprocal` was introduced as an optimization-for-numpy of
`Power(arg,-1)`. At that time `Power` had dtype `float` even when `arg` had
dtype `int`. Necessarily, `Reciprocal` had to inherit this behavior. This
behavior changed in the mean time: the dtype of `Power` is the same as the
dtypes of its arguments and negative integer powers are forbidden, hence
`Power.optimized_for_numpy` always calls `Reciprocal` with a `float` argument.

This patch replaces dtype `float` in `Reciprocal` with the dtype of the
argument of `Reciprocal`. This makes `Reciprocal` - and by extension `Power` -
suitable for complex-valued arguments.