replace evaluable.AsType with Cast
`evaluable.AsType` ensures an array has a specific dtype, applying upcasting if
necessary. If its argument already has the desired dtype, a simplification
removes the cast from the function tree.

There are two problems with this approach. Since a simplification is not
allowed to change the dtype, the simplification that removes the cast happens
if and only if its argument already has the desired dtype. In this case we
could choose to omit the cast from the start and save some steps during
simplification.

Secondly, we might benefit from moving the cast down, because integer
operations are cheaper than floating point operations and floating operations
are cheaper than complex operations. Moving `AsType` down is a bit tedious,
because the dtype of its argument can be anything. For example `Float(a) +
Float(b)` can be simplified to `Float(a + b)` only if `a` and `b` have the same
dtype.

This patch replaces `AsType` with base class `Cast` and `Int`, `Float` and
`Complex` with `BoolToInt`, `IntToFloat` and `FloatToComplex`, casts that go
from one specific dtype to another. Moving `Cast` down is trivial and we don't
introduce no-op casts by design.