add support for complex-valued problems to solvers
Currently, the `solver._redict` helper assumes the problem to be solved is
real-valued. This patch adds a `dtype` parameter to `solver._redict` to inform
the helper which dtype the solution should have. In addition, this patch adds a
helper function `_determine_dtype`, which determines the dtype of the problem.
For now mixed problems are not supported: all targets must have the same dtype.