simplify dispatching of scipy solvers
This patch replaces ScipyMatrix's _solve method by a series of _solve_* methods
that does not require overriding the base implementation. This has two
advantages over the old approach: the selected solver is visible in the output
(e.g. 'gmres' instead of 'scipy') and it allows for better individual control,
used directly to specify gmres's callback_type.