deprecate name arg of mesh, add root arg
This patch replaces `transform.Identifier` with the cheaper (in storage and
comparison) `transform.Index` in `mesh.rectilinear` and `mesh.multipatch`. This
is mostly an internal affair, however, the user had control over a name
component in the `Identifier`. This patch deprecates the name parameter and
adds a `root` parameter as alternative, accepting a `TransformItem` and
defaulting to `Index(ndims, 0)`.