simplify deriv of evaluable.Greater,Less,Equal
The derivatives of `evaluable.Greater`, `evaluable.Less` and `evaluable.Equal`,
subclasses of `Pointwise`, were defined as partial derivatives to its
arguments. While not wrong, `Pointwise._derivative`, in its ignorance, computes
the derivative of the arguments and multiplies this with partial derivatives
defined via `Pointwise.deriv`. Needlessly complex. This patch simply removes
the implementations of `Pointwise.deriv`, which causes `Pointwise._derivative`
to fallback to `Evaluable._derivative`. The latter immediately realises that
the result should be zero based on the dtype of the function. As a side effect
the dtype of the derivative changes from `float` to `bool`.