set default tol=0 for StructureTopology.locate
This patch sets the default value for the structured topology's implementation
of locate so that it matches the baseline implementation. At the default value
of a zero tolerance, the linearity check is disabled as the test is very
unlikely to pass to machine accuracy. This is fully backwards compatible as the
tol argument was mandatory prior to this patch.