support eps in StructuredTopology.locate
This patch adds support for the use of the eps argument instead of or in
addition to the tol argument in the detection of linear geometries in the
structured topology's implementation of locate. If eps==0 (the default) the
implementation is almost backwards compatible, except that the linearity check
measures errors in the infinity norm rather than the euclidean norm.