support disabling linesearch
This patch introduces the syntax `solver.newton(..., linesearch=None)` to
disable linesearching logic. Since the None value was used to support the
legacy arguments `searchrange` and `rebound`, this patch changes the default
linesearch value to "__legacy__" to continue this purpose. The next commit will
add deprecation warnings to the use of these legacy arguments, so that the in
the next release cycle the default value can instead be set to the relevant
NormBased instance.