replace random seed by asymmetric mesh
This patch removes the initial noise that was added to the initial velocity
field to promote bifurcation, and replaces it by 90 degree rotation of the
finite element mesh. By turning the zeroth mesh line vertical, a vertical
asymmetry can be achieved by simply configuring an odd element count, which
should have a similar bifurcating effect.

A side effect of this patch is that the rotation parameter is changed in sign
so that positive rotation is now clockwise, which seems to be the more natural
direction to configure.