fix 1D multitrim
This patch accounts for the situation that a 1D element may have more than two
vertices. This is the case when a line is trimmed for a second time: the mosaic
resulting from the first slice has three edges, one of them empty. A unittest
is added to cover this situation.