change lowerargs to include full transforms seq
This patch changes the `transform_chains` part of the lower args from a list of
`EvaluableTransformChain`s to a tuple of a list of transforms sequences
(`transformseq.Transforms`) and an evaluable index (`Array`). The relevant
operations defined on the `EvaluableTransformChains` are replaced with the
coordinate transformation operations defined in a previous commit.