change lowerargs to include full transforms seq
This patch changes the `transform_chains` part of the lower args from a list of
`EvaluableTransformChain`s to a tuple of a list of transforms sequences
(`transformseq.Transforms`) and an evaluable index (`Array`). The relevant
operations defined on the `EvaluableTransformChains` are now defined as
subclasses of `evaluable.Array`, taking (unevaluable) transforms sequences, the
evaluable index and evaluable coordinates as constructor arguments.