improve MosaicReference.vertices and move to init
This patch makes the MosaicReference object's vertices attribute more efficient
by relying on the property that the new edges have their first vertices in
common with those of the original ones. Furthermore, attribute construction is
moved to the constructor, in anticipation of a future commit that will add
simultaneous tracking of vertex-edge relations.