docs: obtain version of nutils via import
This patches replaces the parsing of the Nutils version number from
`__init__.py` with importing `nutils` and obtaining module attribute `version`.
There are two motivations for this change: Sphinx is going to import `nutils`
anyhow, and in a follow-up commit the line in `__init__.py` defining the
version will be changed, which would necessitate a corresponding change to
`docs/conf.py` without this patch.