replace util.settable by util.set_current
This patch replaces util.settable to util.set_current and rewrites the three
applications: matrix.backend, parallel.maxprox and cache.enable/disable.

Additionally it adds the appdirs dependence, which is used for the new default
location of the function cache. This new location will become relevant in a
future commit when the cache is activated via a context that has no knowledge
of the outrootdir.