support arrays with scale in opposite, mean, jump
This patch adds support to `function.opposite` for arrays that can be cast into
a `function.Array` with a scale.

To support the same type of arrays in `function.mean` and `function.jump` the
explicit cast to `function.Array` (without scale) is removed and we now rely on
the casts performed by `function.opposite`, addition and subtraction. The
downside of this approach is that the casts are performed twice: once in
`function.opposite` and once in the addition or subtraction.