raise some errors as exceptions for easy debugging
This patch raises some errors in util.cli as exceptions, rather than exiting
via sys.exit. The situations where this change is made relate to programming
errors such as invalid signatures, as opposed to the class of errors that are
caused by faulty user input. Raising the former as exceptions is useful for
debugging purposes.