use name_of_main as fallback scriptname
This patch changes the default scriptname from argv[0] to name_of_main. The
reason for this is that argv[0] fails in scenarios such as python -m, python -m
pdb, and other alternative invocations, while name_of_main appears to be fairly
robust.

Instead of changing the default to name_of_main directly, the function default
is set to an empty string, because name_of_main is not reliable during module
initialization. The empty string is substituted by name_of_main in the function
body. Consequently, this changes the meaning of the empty scriptname from the
recently introduced behaviour (omit the script subdirectory) back to the
original behaviour (always make a script subdirectory).