fix gradient of function.Custom
The class `function.Custom` lowers to a class that implements
`evaluable.Array.derivative` in terms of partial derivatives of the loweree:
for all arguments that are instances of `evaluable.Array` with dtype `float` or
`complex` the `function.Custom.partial_derivative` method is called and the
result is dotted with the derivative of the argument to the derivative target.

To increase the performance of the construction of the derivative, arguments
that do not depend on the derivative target are skipped. The dependency test,
however, incorrectly uses the `Evaluable.dependencies` attribute, which does
not contain virtual derivative targets, in particular the root coordinates
(`function._root_derivative_target`). The effect of this bug is that
derivatives of lowered `function.Custom` objects to a virtual derivative target
become zero.

This patch fixes this problem by testing for existence of a derivative target
using the `Evaluable.arguments` attribute and adds a unittest.