add envvar NUTILS_MATRIX_MKL_LIB
In GitHub Actions we've not been able to load MKL with `util.loadlib`, even
when adding the path to `mkl_rt.?.dll` or `libmkl_rt.?.dylib` to `PATH` or
`DYLD_LIBRARY_PATH`. To be able to test MKL and to aid users of the
aformentioned platforms having the same problem, this patch adds the
environment variable `NUTILS_MATRIX_MKL_LIB`, which should be a path to
`mkl.?.dll` on Windows, `libmkl_rt.?.dylib` on MacOS and `libmkl.so.?` on
Linux. If unempty, Nutils tries to load MKL from the provided path. If this
fails, an execption will be raised. If the user provides a path to a different
library, the behavior is undefined.