support interpolation in sample.basis
This patch adds the 'interpolation' argument to sample.basis, with default
value "none" to match the old requirement that the basis is evaluated in the
identical sample. The only other valid value (for now) is "nearest", which
allows evaluation in any sample by selecting the points that are closest to the
target.