enhance test for non-destructive slice
A slice operation returns the empty element in case only a single edge is
changed, or, conversely, the full element in case all but one edges are
unchanged. This patch relaxes this criterion for elements of 3D and up by
extending it to ndims-1 edges. The reasons for this are twofold: 1. the test is
more powerful, resulting in fewer unnecessary mosaic elements, and 2. the new
criterion generalizes to 1D, which is used in a subsequent commit to simplify
the slice method.