make evalf of Evaluable impls static if possible
To make a repeated call to `f` cheaper, where `f = arr.evalf` and `arr` is an
instance of an implementation of `Evaluable`, this patch makes all `evalf`
methods of all implementations of `Evaluable` static if they don't depend on
`self`. For example, with this patch `add.evalf`, where `add` is an instance of
`Add`,  evaluates to `numpy.add` instead of a method call.