enhance function._is_unit_scalar
The _is_unit_scalar method is used to to simplify operations such as
multiplication with unit-valued arguments at the function layer. This patch
enhances its logic in such a way that it tests True for a broader class of
arguments including those that are already cast to _Constant. This prepares the
way for a future commit that will result in more aggressive casting, which
would otherwise disable many of these simplifications.