make Array.cast more efficient
This patch changes the way in which nested data is cast to an Array. The main
reason for the change is that the opportunistic call to _Constant was
potentially expensive, as the call to asarray (via arraydata) could lead to
expensive iteration (casting and stacking results) before concluding that the
resulting object is not actually numerical. To prevent this, we need to be able
to determine a priori whether an object can succesfully be made _Constant.
However, as valid data can take the form of deeply nested data structures, this
test is not trivial. This patch solves the problem by stacking the data in a
depth-first fashion, and conclude by either accepting the result if it is
Array, or otherwise wrapping it in a _Constant. At this stage all nested data
is taken care of by stack, and a cheap test for allowable objects remains.