Avoid unnecessary copy in creation of arraydata
This patch adds copy=None to astype in the arraydata constructor, which used to
always make a copy of the source array just before tobytes would copy it again.
In the new form, the first copy is made only if the source dtype is not the
platform default, e.g. int32 instead of int64.