add optional naxes param to evaluable.unalign
Several simplifications use `unalign` but require the last axes to be
untouched. Currently, if such an axis is uninserted or transposed by `unalign`,
the simplification reinserts and transposes the axis to the end. To aid
simplifications of this kind, this patch adds an `naxes` parameter to
`evaluable.unalign`, which, if not `None`, restricts the number of axes
considered for unalignment.