reimplement Evaluable._stack using inspect
This patch changes the low-level function inspection of Evaluable._stack to a
high level equivalent based on inspect.signature. This fixes the problem that
error messages change in debug mode due ignoring the __signature__ of the debug
wrapper. In the process, the output is also made slightly cleaner by removing
colons and commas.