fix error check in Sampled.evalf
This patch changes the assertion in evaluable.Sampled.evalf to a raised
ValueError. Besides the fact the assert should not be used for failure modes
that are potentially user triggered, the test itself was also faulty as the
shapes of the compared arrays could vary. The latter issue is fixed by
comparing the array shapes prior to their content.