drop assert from _array_bases
`types.frozenarray(x, copy=False)` tries to make `x` and all of its bases
read-only. Bases are typically instances of `numpy.ndarray`, `types.arraydata`,
or `None` if there is no base. `types.frozenarray` asserts that all bases are
of these three types to make sure that we don't miss an opportunity to make a
base read-only.

In a follow-up commit the code for manipulating polynomials, `numeric.poly_*`,
will be replaced by an external module written in Rust. The arrays returned by
this module have a base of type `PySliceContainer` (provided by crate
[numpy](https://crates.io/crates/numpy)). To support making these arrays frozen
as well, this patch drops the assertion at the end of `_array_bases`, thereby
allowing all kinds of array bases.