exclude np-1.24 until accuracy issues are resolved
With the release of Numpy 1.24 on December 18, certain unit tests started
failing in Github actions due to lack of accuracy:

- tests.test_evaluable.check:log2-complex.test_pointwise
- tests.test_evaluable.check:sqrt.test_inv

Unfortunately the issue cannot be reproduced locally, and as such is impossible
to debug. Out of caution, this patch excludes Numpy version 1.24 from the
dependency range until we have more clarity about the situation.