make topologies non-singleton
This patch makes topologies non-singleton by no longer deriving Topology from
types.Singleton. The reason for this is that the singleton nature was never
used, so it added complexity for no reason. By removing it we allow arguments
to be processed in the constructor rather than via apply_annotations. Property
caching is now handled by the cached_property decorator rather than the
__cache__ attribute.