add transform.iscanonical + unit tests
This patch adds the iscanonical function to the transform module to establish
whether a chain of transformations is already in canonical order. This will be
used in a later commit that input arguments are properly preprocessed. The
patch also introduces unit tests for the new function as well the existing
functions canonicalize, uppermost, and promote.