Pull inserted axes through TransformCoords
This patch adds a simplification to the TransformCoords operation that removes
inserted axes from its input coordinates and reapplies them to the result.
Besides improving the efficiency of the transformation itself, this
simplification also exposes insertion structure to any operation following it
such as PolyVal, which can then benefit from a similar simplification.