add simplifications for axes of zero length
This patch adds a trivial simplification that removes function branches with
zero array size. While the logic could exist at a generic level (as it depends
only on the array's shape), InsertAxis is the only remaining operation that can
introduce a zero-length axis (Constant and Take already contain relevant
simplificatons) so this patch focuses the test here.