add util.merge_index_map
`MultipatchTopology.basis_spline()` and `Topology._basis_c0_structured()` both
have an implementation for building a global dofmap given elementwise or
patchwise dofmaps and edge connectivity. This patch adds another implementation
to module `_util`. In a follow-up patch the old implementations will be
replaced by this one for the sake of deduplication.