fix creating a C^0 basis on a multipatch topology
The implementation for merging dofs when creating a C^0 basis on a
`MultipatchTopology` fails (with an assertion error) under certain conditions.
What happens is the following. Given four dofs on four patches, `[0, 1, 2, 3]`,
and the pairings `[0, 2]`, `[1, 3]`and `[0, 3]`, the current implementation
produces the following sequence of pointers:

    init      [0, 1, 2, 3]
    # pair -> pointers
    [0, 2] -> [0, 1, 0, 3]
    [1, 3] -> [0, 1, 0, 1]
    [0, 3] -> [0, 1, 0, 0] # incorrect!

This patch fixes the problem by replacing the implementation for merging
patchwise dofmaps with the recently introduced `_util.merge_index_map()`.