change reference solution of sqrt, log2-complex
This patch changes the reference result of the unit tests for sqrt and
log2-complex to match the expresions in the evalf method of the evaluables.
This fixes tolerance issues on CPUs with AVX-512 extensions, where equivalent
operations like sqrt and power(.., .5) yield slightly different results as of
Numpy version 1.24.