remove apply_annotations from solver module
This patch removes all apply_annotation decorators from the solver module, part
of a process to remove this construct from the entire nutils code base.

Since solver methods are user facing we cannot remove the apply_annotations
decorator by being stricter in what inputs are permissible, but instead we make
dedicated preprocessors for all objects that take care of the required
conversions, before handing over the converted arguments to the actual object
(prefixed with an underscore). The dedicated approach makes it possible to
absorbe the single_or_multiple decorator, as well as process legacy constructs
such as the lhs0 argument, and prefixed arguments such as solveargs.

In changes to the old situation, the preprocessor functions have their optional
arguments made keyword-only, and the lhs0 argument has been restricted to the
single-argument situation.