fix single style solver caching
This patch fixes an issue with the way the single_or_multiple decorator
processed the 'single' situation, which led the solve and solve_withinfo
methods to raise an exception when used with caching unabled due to one of the
arguments being a generator. The _iter unittests are renamed to _cache and
extended to include the solve methods where applicable.