create log context for solve, solve_withinfo
This patch places newton.solve, minimize.solve, etc in an overarching context
that contains separate contexts for its iterates. This structure makes it
easier for interactive log viewers to fold up iterative procedures.

This patch also makes the tol argument mandatory, which should not be a
breaking change as the original default of tol=0 always had to overridden to be
useful.