fix locate for trimmed domain w/ skip_missing=True
This patch fixes the behaviour of SubsetTopology.locate when skip_missing is
set. Formerly it would fail if points that were found on the base topology
(where skip_missing was honoured) but not in the subset. With this patch the
flag instead triggers a post-processing step that prunes the initial sample.