fix error type on parallel locate of missing point
This patch fixes the exception raised by TransformChainsTopology.locate in a
parallel context. By raising LocateError from inside parallel.ctxrange, it
could happen that the exception was emitted by a child process, caught by the
main process as a return status, and reraised as a generic exception. The
modified code communicates failure via shared memory and raises from the main
process instead so that the exception type is reliable.