replace __cache__ by @cached_property, @lru_cache
This patch removes all uses of the __cache__ mechanism, facilitated by the
CacheMeta meta class, and replaces it by the cached_property and lru_cache
decorators which are considered to give better visibility to caching
structures. The CacheMeta class will be removed in a subsequent commit.

NOTE: The only method that is newly uncached as a result of this patch is
TensorReference.get_poly_coeffs, which is a bit tricky to get right because of
the kwargs. Since all other get_poly_coeffs methods are uncached to begin with,
it is unclear whether caching of the particular method is actually relevant.