improve simplification of multiplied additions
This patch improves the simplification of the situation (a + b) * c to a * c +
b * c. To make sure this transformation is only applied when the new form is
cheaper, the simplification used to be bound both a and b being inflated or
diagonalized. This patch extends the criterion to testing for general _multiply
simplifications as well.