fix Constant._inverse
This patch fixes a bug in the simplification of the inverse of a constant,
which failed to move the inverted axes back to their original positions. The
bug was never triggered because the Constant._transpose simplification is
always executed prior. This also implies that no suitable unit test can be
devised to confirm validity of the new implementation, but the existing
check:const test verifies the non-trivial axes if at any point in time the
(questionable) Constant_transpose simpification is removed.