add nutils_hash support for mutable objects
This patch adds support for mutable data. The reasons for this are threefold: 1.
the immutability was already not guaranteed by all suppored objects, namely
dataclasses and objects that define __getnewarg__, 2. dropping the immutability
requirement opens up the possibility to extend support to mutable built-in
types, and 3. the immutability requirement is not important in practice.

Mutable data was formerly explicitly excluded as a security safeguard, similar
to how Python's built-in hash function is not defined te prevent mutable
objects from being used in hash maps. The nutils hash, however, is used mainly
in on-disk caching scenarios, where it serves to fingerprint the input data
only at the moment of a function call. While this leaves the concern that the
method might modify the argument in-place (thus causing a change when a cached
return value causes the function call to be skipped entirely) these are issues
that can be dealt with at the caching level, either by adding immutibility
tests where relevant or by clear agreements that cached functions should be
clear of side effects.