extend tri/hull support for nested multiplications
For a multiplied sample A * B, the .tri implementation added trivial
implementations for the situation that either A or B consisted of only one
point, or A is one-dimensional. This patch extends this logic to any
multiplication chain A0 * A1 * ... with support for all left and right
multiplications of single-point or one-dimensional samples. In particular this
extends support to 3D structured boundaries.