improve out-of-bounds check of numpy.take
This patch adds a special case to the dispatch of numpy.take for the situation
that indices are constant, in which case an out-of-bounds check can be
performed immediately, rather than during lower when the cause of the issue can
be hard to trace back. An additional advantage is a slight simplification of
the function tree by avoiding evaluable.NormDim.