include arguments in evaluable is constant test
The `WithDerivative` evaluable defines a derivative of a wrapped array
to some target. The target is not included in the
`Evaluable.dependencies`, but is included in the `Array.arguments`. The
current implementation of `Evaluable.isconstant` only looks at its
`.dependencies`, not its `.arguments`, giving a false positive if a
constant array is wrapped inside a `WithDerivative`. This patch fixes
the bug by testing for emptyness of `.arguments` in
`Evaluable.isconstant`, this in addition to the original test to be on
the safe side.