remove evaluable Array size attribute
This patch removes the size attribute of evaluable module's Array class, which
was barely used and in one case used incorrectly, and which had an inconsistent
return value, returning the integer 1 if ndims == 0 or an Array otherwise. The
incorrect use was in the simplification of Take, which intended to return zero
in case the indices had an axis of zero length, but which was rendered inactive
due to testing for integer equality. This patch fixes the simplification.