add simplification for multiplied insertaxis
This patch adds the InsertAxis._multiply simplification, replacing the
flop-reducing logic in Multiply._multiply. While matching insertions in nested
multiplications may presently not all be discovered due to the way that
simplifications are organized, the next commit will improve on this so that the
space of nested configurations is properly explored.