include all modules in doctests
This patch changes doctest's test finder to include all files in the nutils
directory, rather than the ones listed in nutils.__all__ which is due to be
removed. The change includes a modified version of DocTestFinder to fix a bug
triggered by the SI module (formerly untested) and for which a fix is pending
for cpython (https://github.com/python/cpython/pull/107716). The patch also
includes a small fix in SI's documentation.