fix faulty assertion: nverts should be nedges
This patch fixes an assertion in the MosaicReference constructor that checks
that the edge transforms of the new edges coinside with the edge transforms of
the edges of the base reference. Here `nedges` was mistaken for `nverts`, which
worked because in dimensions 1 and 2 the two are often equal, but which was
technically incorrect.