compute element volume based on gaus1 i/o centroid
This patch changes the implementation of Element.volume to use degree=1 gauss
quadrature points, rather than _centroid. Though the two are often equal, the
new implementation is robust against the situation that an element has zero
quadrature points. It also avoids the unnecessary computation of the centroid
coordinate which is not used for the volume.