improve early return in Reference.slice
This patch changes the sign test in Reference.slice that results in an early
return if the levelset values are all of the same sign. By focusing this tests
on the vertices used by simplices, rather than all vertices (some of which may
not play a role due to earlier slicing) more cases might be caught here rather
than in the later edge test, which is now restricted to ndims >= 2.