fix parsegmsh in presence of multiple periodicity
This patch fixes parsing of msh files containing of adjacent periodic
boundaries, which caused post processing of the mirrored vertex identities to
fail. A unit test for this problematic meshing scenario will be added with the
cs module in a subsequent commit.

The flawed post processing logic is replaced by _util.merge_index_map, which
receives a new "preserve" argument, False by default, causing the function to
return the index map without remapping indices to a contiguous range.