fix basis of SimplexTopology
The implementations of the degree one std and the bubble bases for the
`SimplexTopology` incorrectly assume that the `simplices` array contains
vertex in `range(nverts)` without holes, and use this array as is as
dofs, thereby introducing zeros in the basis for every dof in
`set(range(simplices.max() + 1)) - set(simplices.flat)`. The boundary of
a `SimplexTopology`, however, simply returns a subset of the volume
`simplices` array, without renumbering.

This bug was introduced by 7e29a283, which removes `_renumber` as
preprocessor for the `simplices` parameter of
the `SimplexTopology` constructor.

To fix this problem this patch adds a `contiguous_simplices` property
and uses thses as dofs. The alternative, requiring constructing a
`SimplexTopology` with contiguous `simplices`, was rejected because this
puts a computational burden on every boundary of a `SimplexTopology`,
even if we don't create a basis.