fix empty boundary of SimplexTopology
The constructor of `SimplexTopology` trips over

    assert numpy.greater(self.simplices[:, 1:], self.simplices[:, :-1]).all()

if there are no simplices. This is triggered when taking the boundary of
a fully periodic mesh.

This patch fixes the problem by returning an `EmptyTopology` when the
boundary is empty. In addition this patch adds a slightly improved error
message when trying to create an empty `SimplexTopology.