swap priority of sum and take, takediag
This patch changes the simplification priority of takediag, take, and sum, from
high to low in that order. The rationale is that takediag is the cheapest
operation, being just a view. Both takediag and (usually) take reduce the work
of a subsequent summation. Summation followed by take or takediag requires more
flops.

This patch furthermore adds two missing simplifications for _takediag.