add evaluable.Holomorphic
A large portion of the `evaluable.Pointwise` subclasses are holomorphic. This
patch adds a subclass `evaluable.Holomorphic` derived from `Pointwise`, with a
default implementation for `return_type` and support for complex derivatives,
and lets the holomorphic pointwise array functions derive from `Holomorphic`.
This patch also removes support for complex derivatives of subclasses of
`Pointwise` (unless overriden).