fix simplification of Equals involving Range
This patch adds a condition to Equals.simplify to avoid that r[:,_] == r[:,_]
is mistaken for r[:,_] == r[_,:], with r a Range and _ a newaxis. Though the
former case of trivial equality is handled first, the bug could be triggered in
case both sides of the equation are constructed in different order and
therefore not identical, causing it to fall through to the second test which
failed to check the insertion axes.