support locate on tensorial topologies
This patch add support for locate on tensorial topologies, by moving the locate
implementation from TransformChainsTopology to the Topology base class. This is
mode possible by adding a new method _lower_args, as well as an implementation
of _Mul._sample. The Sample.zip method received a new optional ndims argument
to facilitate zipping of samples with varying dimensions.