scale graphviz svg down to 2/3 of original size
This patch modifies the width and height parameters of the graphviz generated
svg image in order to scale it down to 2/3 the original size. This avoids
having to scale the image via the browser which may affect other tabs as well.
Sadly it does not seem possible to instruct the dot executable to generate an
image at the desired scale directly. The regular expression that is used
instead is fairly sensitive to the generated xml, but that is acceptable given
that the only risk is to leave the svg at its original scale.