introduce DimensionError
This patch introduces the SI.DimensionError class derived from TypeError, which
is raised instead of TypeError whenever an operation fails due to incompatible
units. In addition to allowing for more specific exception handling in general,
this will be used in particular in a subsequent commit to handle failing dunder
operations.