Revert "replace const Arrays with Constants in optimize"
Commit f914a19b7836498f4ccd81e01fdb7b1e3a57b5ba causes constant evaluable
dependencies of some target to be evaluated and replaced with `Constant`s. This
change was assumed to be penalty free if the target is evaluated at least once.
It turns out that evaluating the constants separately during optimize and doing
one evaluation of the target is more expensive than before the commit, probably
because of duplicate work in `LoopConcatenate`s (because not combined). This
patch reverts the commit.