add graph.TupleNode
`Evaluable`s like `Eig` and `LoopConcatenateCombined` return tuples of arrays.
Dependencies use `ArrayFromTuple` to fetch one item of the tuple. For
readability `ArrayFromTuple` is hidden from the graph. Currently this is done
by looking for a `_node_tuple` method, implemented by `Eig` and
`LoopConcatenateCombined`, which returns a tuple of nodes. This patch
introduces a `graph.TupleNode` with publicly accessible item nodes, which can
be used to merge `Evaluable._node_tuple` into `Evaluable._node`.